backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.6.0
pymongo==4.9.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
pymupdf==1.23.8
python-multipart==0.0.6
google-generativeai==0.3.2
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.3.0
argon2-cffi==25.1.0

firebase-admin==6.5.0
